fix: rename pipewatch/watemark.py -> watermark.py and update imports
The module was named 'watemark.py' (missing 'r'), which mismatched
'watermark_reporter.py' and 'tests/test_watermark.py'. Any external
caller doing 'from pipewatch.watermark import WatermarkChecker' would
hit ImportError.

Changes:
- rename pipewatch/watemark.py -> pipewatch/watermark.py
- update 'from pipewatch.watemark import ...' in watermark_reporter.py
  and tests/test_watermark.py to 'from pipewatch.watermark import ...'
- update module docstring on test_watermark.py

Closes #15
